# implement vectorized evaluation for builtinSqrtSig