--- a/expression/builtin_math_vec.go
+++ b/expression/builtin_math_vec.go
@@ -40,3 +40,25 @@ func (b *builtinLog10Sig) vecEvalReal(input *chunk.Chunk, result *chunk.Column) 
 func (b *builtinLog10Sig) vectorized() bool {
 	return true
 }
+
+func (b *builtinSqrtSig) vecEvalReal(input *chunk.Chunk, result *chunk.Column) error {
+	if err := b.args[0].VecEvalReal(b.ctx, input, result); err != nil {
+		return err
+	}
+	f64s := result.Float64s()
+	for i := 0; i < len(f64s); i++ {
+		if result.IsNull(i) {
+			continue
+		}
+		if f64s[i] < 0 {
+			result.SetNull(i, true)
+		} else {
+			f64s[i] = math.Sqrt(f64s[i])
+		}
+	}
+	return nil
+}
+
+func (b *builtinSqrtSig) vectorized() bool {
+	return true
+}