### What problem does this PR solve?
implement vectorized evaluation for builtinSqrtSig, for [#12105](https://github.com/pingcap/tidb/issues/12105)

### What is changed and how it works?
according to benchmark, about 5 times faster than before:
```
BenchmarkVectorizedBuiltinMathFunc/builtinSqrtSig-VecBuiltinFunc-8                500000              2444 ns/op            0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinMathFunc/builtinSqrtSig-NonVecBuiltinFunc-8             100000             13055 ns/op            0 B/op          0 allocs/op
```

ps:if we don't check ```result.IsNull(i)```in the loop,the method is only 3 times faster than before, so it's better to add those check nulls codes:
```
BenchmarkVectorizedBuiltinMathFunc/builtinSqrtSig-VecBuiltinFunc-8                300000              4281 ns/op            0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinMathFunc/builtinSqrtSig-NonVecBuiltinFunc-8             100000             14611 ns/op            0 B/op          0 allocs/op
```
### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test